requirements.txt
Django==1.10.6
django-rest-auth==0.9.1
djangorestframework==3.6.2
requests==2.13.0
six==1.10.0
